Adds authentication with backend

+++ frontend/src/components/WorkoutDetails.js
@@ -1,12 +1,19 @@
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
 import { formatDistanceToNow } from "date-fns";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext();
+  const { user } = useAuthContext();
 
   const handleClick = async () => {
+    if (!user) return;
+
     const response = await fetch(`/api/workouts/${workout._id}`, {
-      method: "DELETE"
+      method: "DELETE",
+      headers: {
+        "Authorization": `Bearer ${user.token}`
+      }
     })
     const json = await response.json();
 
@@ -20,7 +27,7 @@ const WorkoutDetails = ({ workout }) => {
       <h4>{workout.title}</h4>
       <p><strong>Load (kg): </strong> {workout.load}</p>
       <p><strong>Reps: </strong> {workout.reps}</p>
-      <p>{formatDistanceToNow(new Date(workout.createdAt), {addSuffix: true})}</p>
+      <p>{formatDistanceToNow(new Date(workout.createdAt), { addSuffix: true })}</p>
       <span onClick={handleClick} className="material-symbols-outlined">delete</span>
     </div>
   );
